requirements.txt
# AI Visual Guidance System - Requirements
# Python 3.10+ required
# AMD MI300X GPU with ROCm 7.0+

# =============================================================================
# WEB FRAMEWORK
# =============================================================================
Flask==3.1.2
Flask-SocketIO==5.5.1
Flask-CORS==6.0.1
python-socketio==5.14.2
python-engineio==4.12.3
simple-websocket==1.1.0
bidict==0.23.1

# =============================================================================
# DEEP LEARNING - AMD ROCM OPTIMIZED
# =============================================================================
# PyTorch with ROCm 7.0 support (nightly build for AMD MI300X)
# Install manually first: pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/rocm7.0
# Or skip if already installed
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# =============================================================================
# COMPUTER VISION
# =============================================================================
ultralytics==8.3.221  # YOLOv8 object detection
ultralytics-thop==2.0.17  # Model efficiency metrics
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88  # Headless version for servers

# =============================================================================
# IMAGE PROCESSING & NUMERICAL COMPUTING
# =============================================================================
numpy==2.1.2
Pillow==12.0.0
scipy==1.15.3

# =============================================================================
# DEPTH ESTIMATION - Depth-Anything V2
# =============================================================================
timm==1.0.21  # PyTorch Image Models (required for Depth-Anything V2)

# =============================================================================
# NETWORKING & UTILITIES
# =============================================================================
requests==2.32.5
urllib3==2.5.0
dnspython==2.8.0

# =============================================================================
# GOOGLE GEMINI AI AGENT
# =============================================================================
google-generativeai>=0.8.3  # Gemini SDK with Live API support
websockets>=13.0  # WebSocket client for Gemini Live API
python-dotenv==1.0.1  # Environment variable management

# =============================================================================
# AUDIO PROCESSING
# =============================================================================
# pyaudio>=0.2.14  # Audio I/O for speech capture (OPTIONAL - requires system packages)
soundfile>=0.12.1  # Audio file I/O
# webrtcvad>=2.0.10  # Voice Activity Detection (OPTIONAL - requires python3-dev)

# =============================================================================
# INSTALLATION NOTES
# =============================================================================
# 1. Install Conda/Miniconda first
# 2. Create environment: conda create -n rocmserve python=3.10
# 3. Activate: conda activate rocmserve
# 4. Install ROCm PyTorch FIRST (before this requirements.txt):
#    pip3 install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/rocm7.0
# 5. Then install other packages: pip install -r requirements.txt
#
# For AMD MI300X GPUs:
# - Requires ROCm 7.0+ drivers installed on system
# - Verify GPU: rocm-smi
# - Verify PyTorch ROCm: python -c "import torch; print(torch.cuda.is_available())"

# Development (optional)
# pytest>=7.4.0
# black>=23.0.0
